infra: cleanup
